# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Dreamer AI Solutions API
  description: |
    RESTful API for Dreamer AI Solutions platform.
    
    ## Authentication
    The API uses JWT-based authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited based on user tier:
    - Anonymous: 100 requests/15 minutes
    - Authenticated: 1000 requests/15 minutes
    - Premium: 5000 requests/15 minutes
    - Enterprise: Custom limits
    
    ## Error Responses
    All errors follow RFC 7807 Problem Details format.
  version: 1.0.0
  contact:
    name: Dreamer AI Solutions Support
    email: support@dreamerai.io
    url: https://dreamerai.io/support
  license:
    name: Proprietary
    url: https://dreamerai.io/terms
servers:
  - url: https://api.dreamerai.io/v1
    description: Production server
  - url: https://api-staging.dreamerai.io/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Development server
...
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      tags:
        - Services
      summary: List all available services
      operationId: listServices
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Services list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesListResponse'
  /services/{slug}:
    get:
      tags:
        - Services
      summary: Get service details
      operationId: getService
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ServicesListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Service'
    ServiceResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          $ref: '#/components/schemas/ServiceDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        isActive:
          type: boolean
    ServiceDetail:
      allOf:
        - $ref: '#/components/schemas/Service'
        - type: object
          properties:
            features:
              type: array
              items:
                type: string
            pricing:
              type: object
            createdAt:
              type: string
              format: date-time
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        requestId:
          type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        pagination:
          type: object
          properties:
            page:
              type: integer
            perPage:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasMore:
              type: boolean
        links:
          type: object
          properties:
            first:
              type: string
            last:
              type: string
            next:
              type: string
            prev:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://api.dreamerai.io/errors/not-found
            title: Resource Not Found
            status: 404
            detail: The requested resource was not found
...